% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_read.R
\name{decode_seqqual_compact}
\alias{decode_seqqual_compact}
\title{Decode compact \code{seq} and \code{qual} columns in BamScale output}
\usage{
decode_seqqual_compact(
  x,
  seq_col = "seq",
  qual_col = "qual",
  qwidth_col = "qwidth"
)
}
\arguments{
\item{x}{A \code{data.frame}, \code{S4Vectors::DataFrame}, or list-like object
containing compact BamScale \code{seq} and/or \code{qual} columns.}

\item{seq_col}{Name of the compact sequence column.}

\item{qual_col}{Name of the compact quality column.}

\item{qwidth_col}{Name of the read-width column used to decode compact
sequence bytes.}
}
\value{
\code{x} with compact \code{seq} and/or \code{qual} columns replaced by decoded
character vectors. The input class is preserved.
}
\description{
Convenience wrapper for converting a compact \code{bam_read()} result back to
ordinary sequence and quality strings.
}
\examples{
x <- data.frame(qwidth = 4L)
x$seq <- I(list(as.raw(c(0x12, 0x48))))
x$qual <- I(list(as.raw(c(0L, 1L, 2L, 3L))))
decode_seqqual_compact(x)
}
\seealso{
\code{\link[=decode_compact_seq]{decode_compact_seq()}}, \code{\link[=decode_compact_qual]{decode_compact_qual()}}, \code{\link[=bam_read]{bam_read()}}
}
