% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BamScale-package}
\alias{BamScale}
\alias{BamScale-package}
\title{BamScale: Bioconductor-Friendly Multithreaded BAM Processing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Multithreaded sequential BAM processing built on top of the ompBAM C++ engine. BamScale provides user-friendly BAM read and scan interfaces designed for compatibility with existing Bioconductor workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cparsania.github.io/BamScale/}
  \item Report bugs at \url{https://github.com/cparsania/BamScale/issues}
}

}
\author{
\strong{Maintainer}: Chirag Parsania \email{chirag.parsania@gmail.com}

}
\keyword{internal}
