% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BamScale-package}
\alias{BamScale}
\alias{BamScale-package}
\title{BamScale: Bioconductor-Friendly Multithreaded BAM Processing}
\description{
Multithreaded sequential BAM processing built on top of the ompBAM C++ engine. BamScale provides user-friendly BAM read and scan interfaces designed for compatibility with existing Bioconductor workflows.
}
\author{
\strong{Maintainer}: Chirag Patel \email{chirag@example.com}

}
\keyword{internal}
